+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Version 0.7.0 / 2018-Oct-12 12:04
+* Update golang to v1.11
+* Switch scripts to develop branch
+* Swithc demo apps to master branch
+* Update frontend to 0.9.2
+* Update backend to 1.0.4
+* Update Block Explorer to 0.2.6
+
 ## Version 0.6.15 / 2018-Aug-20 05:40
 * Add dump-be-all and restore-be-all command to dump/restore backends logs, data directories and databases
 * Fix priv-keys command